+++ Tetris++/Tetris++/Player.h
@@ -17,6 +17,6 @@ private:
 	uint16_t m_score;
 	static uint16_t m_highScore;
 	static uint16_t m_averageScore;
-	static int m_scoreCount;
+	static uint8_t m_scoreCount;
 };
